requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.11.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.0
transformers==4.36.0
requests==2.31.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
numpy==1.24.3
scipy==1.11.4
openai>=1.3.7
firebase-admin==6.2.0
